+++ jlle/scaffold/templates/docs/source/conf.py
@@ -53,7 +53,7 @@ extensions = ['sphinx.ext.autodoc', 'sphinx.ext.todo', 'sphinx.ext.coverage',
 intersphinx_mapping = {'python': ('http://docs.python.org/3', None)}
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+#templates_path = ['_templates']
 
 # The suffix of source filenames.
 source_suffix = '.rst'
@@ -140,7 +140,12 @@ for mod_name in MOCK_MODULES:
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'pyramid'
+
+on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
+if not on_rtd:  # only import and set the theme if we're building docs locally
+    import sphinx_rtd_theme
+    html_theme = "sphinx_rtd_theme"
+    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -181,14 +186,14 @@ html_theme = 'pyramid'
 
 # Custom sidebar templates, maps document names to template names.
 #html_sidebars = {}
-html_sidebars = {
-   '**': [
-          'links.html',
-          'relations.html',
-          'localtoc.html',
-          'searchbox.html',
-          'sourcelink.html'],
-}
+#html_sidebars = {
+#   '**': [
+#          'links.html',
+#          'relations.html',
+#          'localtoc.html',
+#          'searchbox.html',
+#          'sourcelink.html'],
+#}
 
 # Additional templates that should be rendered to pages, maps page names to
 # template names.